+++ source/include/Ostap/Params.h
@@ -25,6 +25,8 @@ namespace Ostap
     class LegendreSum4 ;
     class ChebyshevSum ;
     class Bernstein    ;
+    class Bernstein2D  ;
+    class Bernstein3D  ;
     // ========================================================================
   } //                                         The end of namespace Ostap::Math
   // ==========================================================================
@@ -364,7 +366,125 @@ namespace Ostap
       const std::string&         selection  , 
       const unsigned long        first      =  0 ,
       const unsigned long        last       = std::numeric_limits<unsigned long>::max() ) ;
+    // ==================================================================================
+    /** fill BErnstein with data from the Tree 
+     *  @see Ostap::Math::Bernstein2D
+     *  @see Ostap::Math::Bernstein2D::fill
+     *  @param tree        (INPUT)  the input tree 
+     *  @param sum         (UPDATE) the parameterization object 
+     *  @param xexpression (INPUT)  x-expression to be parameterized
+     *  @param yexpression (INPUT)  y-expression to be parameterized
+     *  @param first       (INPUT)  the first event in Tree 
+     *  @param last        (INPUT)  the last  event in Tree 
+     *  @return number of events used in parameterization 
+     *  @code
+     *  Tree*  tree = ...
+     *  Bernsteinn2D s ( 5 , 3 , -1 , 1 , -2 , 2 ) ;
+     *  DataParam::parameterize ( tree , s , "x" , "y/z") ;
+     *  @endcode
+     *  @attention it is less CPU efficient than Ostap::Math::LegendreSum2 
+     *  @see Ostap::Math::LegendreSum2 
+     *  @see Ostap::Math::LegendreSum2::fill
+     *  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+     *  @date   2022-12-26
+     */ 
+    static unsigned long parameterize 
+    ( TTree*                     tree        , 
+      Ostap::Math::Bernstein2D&  sum         , 
+      const std::string&         xexpression , 
+      const std::string&         yexpression , 
+      const unsigned long        first       =  0 ,
+      const unsigned long        last        = std::numeric_limits<unsigned long>::max() ) ;
+    // ========================================================================
+    /** fill BErnstein with data from the Tree 
+     *  @see Ostap::Math::Bernstein2D
+     *  @see Ostap::Math::Bernstein2D::fill
+     *  @param tree        (INPUT)  the input tree 
+     *  @param sum         (UPDATE) the parameterization object 
+     *  @param xexpression (INPUT)  x-expression to be parameterized
+     *  @param yexpression (INPUT)  y-expression to be parameterized
+     *  @param selection   (INPUT)  selection/weight to be used 
+     *  @param first       (INPUT)  the first event in Tree 
+     *  @param last        (INPUT)  the last  event in Tree 
+     *  @return  sum of weigths  used in parameterization
+     *  @code
+     *  Tree*  tree = ...
+     *  Bernstein2D s ( 5 , 2 ,  -1 , 1 , -4  , -5 ) ;
+     *  DataParam::parameterize ( tree , s , "x"  , "z" , "y>10" ) ;
+     *  @endcode
+     *  @attention it is less CPU efficient than Ostap::Math::LegendreSum2 
+     *  @see Ostap::Math::LegendreSum2 
+     *  @see Ostap::Math::LegendreSum2::fill
+     *  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+     *  @date   2022-12-26
+     */ 
+    static double parameterize 
+    ( TTree*                     tree        ,  
+      Ostap::Math::Bernstein2D&  sum         , 
+      const std::string&         xexpression , 
+      const std::string&         yexpression , 
+      const std::string&         selection   , 
+      const unsigned long        first       =  0 ,
+      const unsigned long        last        = std::numeric_limits<unsigned long>::max() ) ;
     // =======================================================================
+    /** fill Bernstein with data from the Tree 
+     *  @see Ostap::Math::Bernstein3D
+     *  @see Ostap::Math::Bernstein3D::fill
+     *  @param tree        (INPUT)  the input tree 
+     *  @param sum         (UPDATE) the parameterization object 
+     *  @param xexpression (INPUT)  x-expression to be parameterized
+     *  @param yexpression (INPUT)  y-expression to be parameterized
+     *  @param zexpression (INPUT)  z-expression to be parameterized
+     *  @param first       (INPUT)  the first event in Tree 
+     *  @param last        (INPUT)  the last  event in Tree 
+     *  @return number of events used in parameterization 
+     *  @code
+     *  Tree*  tree = ...
+     *  Bernstein3D s ( 5 , 3 , 2 , -1 , 1 , -2 , 2 , 0 , 4 ) ;
+     *  DataParam::parameterize ( tree , s , "x" , "y" , "y/z") ;
+     *  @endcode
+     *  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+     *  @date   2019-07-3
+     */ 
+    static unsigned long parameterize 
+    ( TTree*                     tree        , 
+      Ostap::Math::Bernstein3D&  sum         , 
+      const std::string&         xexpression , 
+      const std::string&         yexpression , 
+      const std::string&         zexpression , 
+      const unsigned long        first       =  0 ,
+      const unsigned long        last        = std::numeric_limits<unsigned long>::max() ) ;
+    // ========================================================================
+    /** fill Bernstein with data from the Tree 
+     *  @see Ostap::Math::Bernstein3D
+     *  @see Ostap::Math::Bernstein3D::fill
+     *  @param tree        (INPUT)  the input tree 
+     *  @param sum         (UPDATE) the parameterization object 
+     *  @param xexpression (INPUT)  x-expression to be parameterized
+     *  @param yexpression (INPUT)  y-expression to be parameterized
+     *  @param zexpression (INPUT)  z-expression to be parameterized
+     *  @param selection   (INPUT)  selection/weight to be used 
+     *  @param first       (INPUT)  the first event in Tree 
+     *  @param last        (INPUT)  the last  event in Tree 
+     *  @return  sum of weigths  used in parameterization
+     *  @code
+     *  Tree*  tree = ...
+     *  Bernstein3D s ( 5 , 3 , 2 , -1 , 1 , -2 , 2 , 0 , 4 ) ;
+     *  DataParam::parameterize ( tree , s , "x"  , "y" , "z" , "t>10" ) ;
+     *  @endcode
+     *  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+     *  @date   2019-07-3
+     */ 
+    static double parameterize 
+    ( TTree*                     tree        ,  
+      Ostap::Math::Bernstein3D&  sum         , 
+      const std::string&         xexpression , 
+      const std::string&         yexpression , 
+      const std::string&         zexpression , 
+      const std::string&         selection   , 
+      const unsigned long        first       =  0 ,
+      const unsigned long        last        = std::numeric_limits<unsigned long>::max() ) ;
+    // ========================================================================
   } ; //                                      The end of class Ostap::DataParam 
   // ==========================================================================
 } //                                                 The end of namespace Ostap